[JBIDE-25282] - Spring Boot M2 configurator
- Code review
