+++ maven/features/org.jboss.tools.maven.springboot.feature/feature.properties
@@ -1,5 +1,5 @@
 ###############################################################################
-# Copyright (c) 2010-2017 Red Hat, Inc. and others.
+# Copyright (c) 2018 Red Hat, Inc. and others.
 # All rights reserved. This program and the accompanying materials 
 # are made available under the terms of the Eclipse Public License v1.0
 # which accompanies this distribution, and is available at
