+++ maven/plugins/org.jboss.tools.maven.springboot/META-INF/MANIFEST.MF
@@ -12,7 +12,6 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.jst.j2ee.core,
  org.eclipse.jst.j2ee,
  org.eclipse.wst.common.emfworkbench.integration,
- org.jboss.tools.cdi.core,
  org.eclipse.m2e.wtp;bundle-version="1.0.0";resolution:=optional
 Bundle-Localization: plugin
 Bundle-ActivationPolicy: lazy
