+++ maven/plugins/org.jboss.tools.maven.ui/src/org/jboss/tools/maven/ui/Activator.java
@@ -85,7 +85,7 @@ public class Activator extends AbstractUIPlugin {
 
 	public static final boolean CONFIGURE_ARQUILLIAN_VALUE = true;
 	
-	public static final String CONFIGURE_SPRING_BOOT = "configureSpringBoot";
+	public static final String CONFIGURE_SPRING_BOOT = "configureSpringBoot"; //$NON-NLS-1$
 	
 	public static final boolean CONFIGURE_SPRING_BOOT_VALUE = true;
