--- maven/tests/org.jboss.tools.maven.configurators.tests/projects/springboot/jar/src/main/java/org/jboss/tools/maven/tests/springbootjar/SpringbootJarApplication.java
@@ -1,12 +1,0 @@
-package org.jboss.tools.maven.tests.springbootjar;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringbootJarApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringbootJarApplication.class, args);
-	}
-}
