+++ maven/tests/org.jboss.tools.maven.configurators.tests/projects/springboot/war/src/main/java/org/jboss/tools/maven/tests/springbootwar/SpringBootWarApplication.java
@@ -10,14 +10,13 @@
  ************************************************************************************/
 package org.jboss.tools.maven.tests.springbootwar;
 
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.support.SpringBootServletInitializer;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-public class ServletInitializer extends SpringBootServletInitializer {
+@SpringBootApplication
+public class SpringBootWarApplication {
 
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(SpringBootWarApplication.class);
+	public static void main(String[] args) {
+		SpringApplication.run(SpringBootWarApplication.class, args);
 	}
-
 }
