--- maven/tests/org.jboss.tools.maven.configurators.tests/projects/springboot/war/src/main/java/org/jboss/tools/maven/tests/springbootwar/SpringbootWarApplication.java
@@ -1,12 +1,0 @@
-package org.jboss.tools.maven.tests.springbootwar;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringbootWarApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringbootWarApplication.class, args);
-	}
-}
