+++ maven/tests/org.jboss.tools.maven.configurators.tests/projects/springboot/war/src/test/java/org/jboss/tools/maven/tests/springbootwar/SpringBootWarApplicationTests.java
@@ -10,14 +10,17 @@
  ************************************************************************************/
 package org.jboss.tools.maven.tests.springbootwar;
 
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.support.SpringBootServletInitializer;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
 
-public class ServletInitializer extends SpringBootServletInitializer {
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class SpringBootWarApplicationTests {
 
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(SpringBootWarApplication.class);
+	@Test
+	public void contextLoads() {
 	}
 
 }
