--- maven/tests/org.jboss.tools.maven.configurators.tests/projects/springboot/war/src/test/java/org/jboss/tools/maven/tests/springbootwar/SpringbootWarApplicationTests.java
@@ -1,16 +1,0 @@
-package org.jboss.tools.maven.tests.springbootwar;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class SpringbootWarApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
